requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
pydantic==2.7.0
alembic==1.13.1
psycopg2-binary==2.9.9
python-multipart==0.0.9
python-jose==3.3.0
fastapi-mail==1.4.1
bcrypt==4.1.2
SQLAlchemy==2.0.29
python-dotenv==1.0.1
passlib
fastapi-pagination==0.12.23
fastapi[all]